+++ ntex/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## [0.3.12] - 2021-03-18
 
-* http: add per request callback for http1
+* http: add per request handler service for http1
 
 ## [0.3.11] - 2021-03-16